+++ pkg/dns/data.go
@@ -1,0 +1,75 @@
+package dns
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"net/http"
+	"net/url"
+)
+
+type (
+	// Data contains operations available on Data resources.
+	Data interface {
+		// ListGroups returns group list associated with particular user
+		//
+		// See: https://techdocs.akamai.com/edge-dns/reference/get-data-groups
+		ListGroups(context.Context, ListGroupRequest) (*ListGroupResponse, error)
+	}
+
+	// ListGroupResponse lists the groups accessible to the current user
+	ListGroupResponse struct {
+		Groups []Group `json:"groups"`
+	}
+
+	// ListGroupRequest is a request struct
+	ListGroupRequest struct {
+		GroupId string
+	}
+
+	// Group contain the information of the particular group
+	Group struct {
+		GroupID     int      `json:"groupId"`
+		GroupName   string   `json:"groupName"`
+		ContractIDs []string `json:"contractIds"`
+		Permissions []string `json:"permissions"`
+	}
+)
+
+var (
+	// ErrListGroups is returned in case an error occurs on ListGroups operation
+	ErrListGroups = errors.New("list groups")
+)
+
+func (p *dns) ListGroups(ctx context.Context, params ListGroupRequest) (*ListGroupResponse, error) {
+	logger := p.Log(ctx)
+	logger.Debug("ListGroups")
+
+	uri, err := url.Parse("/config-dns/v2/data/groups/")
+	if err != nil {
+		return nil, fmt.Errorf("%w: failed to parse url: %s", ErrListGroups, err)
+	}
+
+	q := uri.Query()
+	if params.GroupId != "" {
+		q.Add("gid", params.GroupId)
+	}
+	uri.RawQuery = q.Encode()
+
+	req, err := http.NewRequestWithContext(ctx, http.MethodGet, uri.String(), nil)
+	if err != nil {
+		return nil, fmt.Errorf("failed to list listZoneGroups request: %w", err)
+	}
+
+	var groups ListGroupResponse
+	resp, err := p.Exec(req, &groups)
+	if err != nil {
+		return nil, fmt.Errorf("ListZoneGroups request failed: %w", err)
+	}
+
+	if resp.StatusCode != http.StatusOK {
+		return nil, p.Error(resp)
+	}
+
+	return &groups, nil
+}